Improving the Game
I changed the gravity, and jump distance.
I changed the interval so that the objects spawn at a slower rate.
I changed the time to run the game faster.
I made the myGamePiece smaller so that the game becomes easier.  The myGamePiece has a square hitbox, so do not judge by the image if you get out, it is the square hitbox that has touched the obstacle.